+++ src/cppREST/SessionManager.cpp
@@ -2,8 +2,7 @@
 #include "Helper.h"
 
 SessionManager::SessionManager()
-	: backup_file_(Helper::openFileForWriting(ServerHelper::getSessionBackupFileName(), false, true))
-	, session_store_()
+    : session_store_()
 	, current_client_info_()
 	, current_notification_()
 {	
@@ -15,70 +14,32 @@ SessionManager& SessionManager::instance()
 	return session_manager;
 }
 
-void SessionManager::rewriteFile()
-{
-	instance().mutex_.lock();
-    instance().backup_file_.data()->close();
-
-    if (instance().backup_file_.data()->remove())
-    {
-        Log::info("Session backup file has been removed");
-        instance().backup_file_ = Helper::openFileForWriting(ServerHelper::getSessionBackupFileName(), false, true);
-
-        QMapIterator<QString, Session> i(instance().session_store_);
-        QTextStream out(instance().backup_file_.data());
-
-        while (i.hasNext())
-        {
-            i.next();
-            if (i.value().user_id > 0)
-            {
-                out << i.key() << "\t" << i.value().user_id << "\t" << i.value().user_login << "\t" << i.value().user_name << "\t" << i.value().login_time.toSecsSinceEpoch() << "\t" << i.value().is_for_db_only << "\n";
-            }
-        }
-
-    }
-    else
-    {
-        Log::error("Could not remove the session backup file");
-    }
-
-	instance().mutex_.unlock();
-}
-
-void SessionManager::saveSessionToFile(QString id, Session in)
-{
-	QTextStream out(instance().backup_file_.data());
-	if (!in.isEmpty())
-    {
-        out << id << "\t" << in.user_id << "\t" << in.user_login << "\t" << in.user_name << "\t" << in.login_time.toSecsSinceEpoch() << "\t" << in.is_for_db_only << "\n";
-	}
-}
-
 void SessionManager::restoreFromFile()
 {
 	if (QFile(ServerHelper::getSessionBackupFileName()).exists())
-	{		
-		int restored_items = 0;
-		if (instance().backup_file_.data()->isOpen()) instance().backup_file_.data()->close();
-		instance().backup_file_ = Helper::openFileForReading(ServerHelper::getSessionBackupFileName());
-		while(!instance().backup_file_.data()->atEnd())
+    {
+        qint64 restored_items = 0;
+
+        QSharedPointer<QFile> backup_file = Helper::openFileForReading(ServerHelper::getSessionBackupFileName());
+        while(!backup_file.data()->atEnd())
 		{
-			QString line = instance().backup_file_.data()->readLine();
+            QString line = backup_file.data()->readLine();
 			if(line.isEmpty()) break;
 
 			QList<QString> line_list = line.split("\t");
             if (line_list.count() > 5)
 			{
 				bool ok;
-				restored_items++;
-                addNewSession(line_list[0], Session(line_list[1].toInt(), line_list[2], line_list[3], QDateTime::fromSecsSinceEpoch(line_list[4].toLongLong(&ok,10)), line_list[5].toInt()), false);
+                Session current_session = Session(line_list[1].toInt(), line_list[2], line_list[3], QDateTime::fromSecsSinceEpoch(line_list[4].toLongLong(&ok,10)), line_list[5].toInt());
+                if (!isSessionExpired(current_session))
+                {
+                    addNewSession(line_list[0], current_session);
+                    restored_items++;
+                }
 			}
 		}
 		Log::info("Number of restored sessions: " + QString::number(restored_items));
-		instance().backup_file_.data()->close();
-
-        removeExpiredSessions();
+        backup_file.data()->close();
 	}
 	else
 	{
@@ -86,11 +47,10 @@ void SessionManager::restoreFromFile()
 	}
 }
 
-void SessionManager::addNewSession(QString id, Session in, bool save_to_file)
+void SessionManager::addNewSession(QString id, Session in)
 {
 	instance().mutex_.lock();
-	instance().session_store_.insert(id, in);
-	if (save_to_file) saveSessionToFile(id, in);
+    instance().session_store_.insert(id, in);
 	instance().mutex_.unlock();
 }
 
@@ -123,7 +83,6 @@ bool SessionManager::isSessionExpired(Session in)
 	// Session lifetime in seconds
 	qint64 valid_period = ServerHelper::getNumSettingsValue("session_duration");
 	if (valid_period == 0) valid_period = DEFAULT_VALID_PERIOD; // default value, if not set in the config
-
 	if (in.login_time.addSecs(valid_period).toSecsSinceEpoch() < QDateTime::currentDateTime().toSecsSinceEpoch())
 	{
 		return true;
@@ -150,27 +109,31 @@ bool SessionManager::isTokenReal(QString token)
 	return false;
 }
 
-void SessionManager::removeExpiredSessions()
+QMap<QString, Session> SessionManager::removeExpiredSessions()
 {
-	QList<QString> to_be_removed {};
+    QMap<QString, Session> tmp_storage = {};
+    QList<QString> to_be_removed {};
+
+    Log::info("Starting to cleanup sessions");
+    instance().mutex_.lock();
 	QMapIterator<QString, Session> i(instance().session_store_);
-	while (i.hasNext()) {
+    while (i.hasNext())
+    {
 		i.next();
 		if (isSessionExpired(i.value()))
 		{
 			to_be_removed.append(i.key());
 		}
 	}
+	for (int i = 0; i < to_be_removed.count(); ++i)
+    {
+        instance().session_store_.remove(to_be_removed[i]);
+	}
+    tmp_storage = instance().session_store_;
+    instance().mutex_.unlock();
 
     Log::info("Number of removed sessions: " + QString::number(to_be_removed.length()));
-
-	for (int i = 0; i < to_be_removed.count(); ++i)
-	{
-		removeSession(to_be_removed[i]);
-	}
-
-    //wipe the backup file out and write active sessions
-    rewriteFile();
+    return tmp_storage;
 }
 
 ClientInfo SessionManager::getCurrentClientInfo()